Started work on the calendar
The calendar page now has columns for each of the days of the week located above the calendar nodes. Above that are two button (eventually for changing month) with the months current name in between them.

The calendar page now has the starting point for the months days begin on the correct section (Monday, Tuesday etc...).
For instance. The current month begins on a Friday, therefore, the node for the first day of this month will be 4 ahead where the Friday column is.