A working calendar
The calendar will now switch between month provided the user click a button and will have the correct day count, weekday start point, year, and year. It will also take leap years into consideration and will automatically start on the current date.

The day will not be selected by default but will instead wait for user input. The right side of the component will soon be changed to show a message if no date is selected.