Calendar additions
Made it so when changing the month and/ or year the currently selected date on the right section is not effected and will remain unchanged until the user clicks another date.

Added a message on the right side and removed the button & current date until the user clicks on a date which will remove the message and replace it with the button & current date.