Booking start & calendar changes
Added code to the booking page so everything is better placed and changed the way typing into text boxes works so that several things can be done when it is being done using listeners.

Added ability to load data from another file. This file will eventually be used to fetch data from the database. The calendar will also print dummy data on the right section for users to view fake bookings.