Calendar changes
Another section has been added to the bottom of the calendar. The intention is to integrate the booking page with the calendar page.

The fake bookings has also been added to a scroll section. However, it has been commented out for now because it breaks the layout when the tabbing system is used.