Bookings additions
An addition has been completed and one is well on the way to being done:
1) Added a search function that will go through several patient records (name, number etc...) and show records matching the text field.

2) Have begun work on the tooth issues selection page. So far you are able to view a 3 by 3 grid and select sections of the grid.
When a section is selected, it will highlight grey. Selecting it again will set it back to normal.
A button has been added to create an additional grid which you can also alter via selection.
Two additional button were added to switch between the grids that were created.

Another small alteration made was having the booking page send the user to the booking dental chart selection for now. This is due to the booking page now being part of the calendar rather than it's own page.