bookings dental chart additions
Added the ability to remove the dental chart currently being viewed. Doing so will default the user back to the first dental chart and reorganise the array holding the charts info.

Added the ability to select a procedure in the dental chart page rather than the booking page. Also moved the patient notes to this location also.

The notes, procedure, and dental charts now successfully save and change in real time.

Added several comments to the booking dental chart code to allow for better understanding and cleaned the code up substantially.